[EI-254] [web] feat: Add select target indicator step to custom forecast indicator create dialog menu
